requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
boto3
python-dotenv
pymongo
redis
requests
